+++ packages/dara-core/js/shared/context/index.tsx
@@ -1,4 +1,3 @@
-export { default as AuthCtx } from '../../auth/auth-context';
 export { default as DirectionCtx } from './direction-context';
 export { default as ImportersCtx } from './importers-context';
 export { default as GlobalTaskProvider, useTaskContext } from './global-task-context';