+++ packages/dara-core/changelog.md
@@ -2,6 +2,12 @@
 title: Changelog
 ---
 
+## NEXT
+
+-   Added support for seamless token refresh mechanism. A provided auth config can be configured to support token refresh by:
+    -   implement the `refresh_token` method to sign a new token, reusing the previous session_id for continuity
+    -   adding some mechanism to set a `dara_refresh_token` cookie, e.g. via custom `components_config` and endpoints such as `/sso-callback` for SSO auth
+
 ## 1.13.1
 
 -   Fixed an issue where data passed to `Table` component within a column of `dtype` `object` did not display correctly for datetime values.